aether-scout/backend/requirements.txt
fastapi==0.111.0
motor==3.4.0
pymongo==4.7.2
python-dotenv==1.0.1
httpx==0.27.0
websockets==12.0
groq==0.9.0
pydantic==2.7.1
uvicorn==0.30.1
ultralytics==8.2.0
opencv-python-headless==4.9.0.80
